% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyalgorithms-package.R
\docType{package}
\name{easyalgorithms-package}
\alias{easyalgorithms}
\alias{easyalgorithms-package}
\title{easyalgorithms: Euclidean and Dijkstra algorithm}
\description{
Contains the algorithms for Euclidean and Dijkstra. These are
    separed into two function.
}
\section{easyalgorithms}{
 A package containing two functions.
First one is \code{\link{euclidean}}.

Second one is \code{\link{dijkstra}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OLGJ/easyalgorithms}
  \item Report bugs at \url{https://github.com/OLGJ/easyalgorithms/issues}
}

}
\author{
Olof Josefsson
}
\keyword{internal}
