% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A numeric input.}

\item{y}{A numeric input.}
}
\value{
The Euclidean of \code{x} and \code{y}.
}
\description{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
