% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki graph sample data.}
\format{
A dataframe with 3 columns.
\describe{
  \item{v1}{From node, numeric.}
  \item{v2}{To node, numeric}
  \item{w}{Associated weight, numeric}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Graph#Mathematics}
}
\usage{
wiki_graph
}
\description{
Sample data from wikipedia page.
}
\details{
A dataset containing 3 columns with nodes, their connection and their
associated weight.
}
\keyword{datasets}
