% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A datafram containing 3 columns ("v1", "v2", "w").}

\item{init_node}{A numeric scalar.}
}
\value{
The dijkstra of \code{graph} and \code{init_node}.
}
\description{

}
\examples{
dijkstra(graph, init_node)
}
